move socket connection to separate file
